refactor: delivery - 캐싱 읽기 전략#138
